Update hardware to the WiModem232 module
Rather than continue making my own board based on the Pololu 23201a,
adapt WiFiPPP to the hardware in the ESP8266-based WiModem232.

This device has an ESP8266-12F module on a custom PCB with DB25
connector, along with an optional I2C-based SSD1306 monochrome OLED
screen (though it has separate areas on it that draw in blue or
yellow), and a NeoPixel 5050 multi-color LED with an integrated
WS2811 controller.

So far I have the screen and NeoPixel working, as well as basic
RS232 input and output with the DTR line mapped.  I still need to
map out the remaining RTS, CTS, DSR, and DCD lines.