AT handling: Do DTR processing as advisory, not required
Just remember if we had DTR during the call and if it drops, hangup,
but don't require DTR to talk.

If we get non-ASCII garbage at the start of a connection, assume
it's a new connection (without DTR) sending stuff to us at the wrong
baud rate and perform an autobaud.

Shrink cur/lastcmd buffers to save a bit of memory.